add callback tests for SAP
Also remove custom handling for 'send-sms-code' reference. This looks
like it was part of some old UK code when the system only sent out SMS
codes and verify emails. This commit shows the reference being removed
from the callers, but it looks like it was never removed in the tests
and in the processing function.